// frontend/src/app/api/generated/model/patient-payment-dto.ts
/**
 * Zana HMIS API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MoneyDto } from './money-dto';


export interface PatientPaymentDto { 
    uid?: string;
    patientUid?: string;
    amount?: MoneyDto;
    paymentType?: PatientPaymentDto.PaymentTypeEnum;
    status?: string;
    businessDayUid?: string;
}
export namespace PatientPaymentDto {
    export const PaymentTypeEnum = {
        Cash: 'CASH',
        Insurance: 'INSURANCE'
    } as const;
    export type PaymentTypeEnum = typeof PaymentTypeEnum[keyof typeof PaymentTypeEnum];
}