// frontend/src/app/api/generated/model/record-payment-request.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MoneyDto } from './money-dto';


export interface RecordPaymentRequest { 
    billUids: Array<string>;
    tenderedTotal: MoneyDto;
    paymentMode: RecordPaymentRequest.PaymentModeEnum;
}
export namespace RecordPaymentRequest {
    export const PaymentModeEnum = {
        Cash: 'CASH',
        Insurance: 'INSURANCE'
    } as const;
    export type PaymentModeEnum = typeof PaymentModeEnum[keyof typeof PaymentModeEnum];
}